refactor textinputlayout matchers
